Create solve.txt
BRIEFING
We have managed to get the login details for the staff portal at Lost Travel: We've come across a page from a book at https://cgames-we02.allyourbases.co where we believe there's extra information hidden., however we still cant seem to get in. Take a look and see if you can find a way in.